restructure dashboard
