+++ packages/admin/src/app/(dashboard)/packages/actions/index.ts
@@ -4,7 +4,7 @@ import path from 'path';
 
 import { config } from '@/lib/framework-utils';
 
-import { PackageService } from '../../../service/service.package';
+import { PackageService } from '@/service/service.package';
 
 export async function isPackageInstalled({ packageName }: { packageName: string }) {
   const packageJsonPath = path.join(process.env.APP_DIR || process.cwd(), 'package.json');
