+++ packages/admin/src/app/(dashboard)/records/[integration]/[entityType]/client-layout.tsx
