+++ packages/admin/src/app/(dashboard)/records/[integration]/page.tsx
