+++ packages/admin/src/app/(onboarding)/page.tsx
@@ -1,0 +1,7 @@
+import React from 'react';
+
+const Index = () => {
+  return <div>Index</div>;
+};
+
+export default Index;
