+++ packages/admin/src/app/layout.tsx
@@ -2,12 +2,9 @@ import { Inter } from 'next/font/google';
 
 import { Toaster } from '@/components/ui/sonner';
 
-import { framework } from '@/lib/framework-utils';
-
 import 'react-day-picker/style.css';
 
 import './globals.css';
-import AdminLayout from './layouts/admin-layout';
 
 const inter = Inter({ subsets: ['latin'] });
 
@@ -16,13 +13,11 @@ export default async function RootLayout({
 }: Readonly<{
   children: React.ReactNode;
 }>) {
-  const integrations = framework?.availableIntegrations() || [];
-
   return (
     <html lang="en">
       <body className={`dark ${inter.className}`}>
         <Toaster position="bottom-right" />
-        <AdminLayout integrations={integrations}>{children}</AdminLayout>
+        {children}
       </body>
     </html>
   );
