--- packages/admin/src/app/page.tsx
@@ -1,3 +1,0 @@
-export default function Home() {
-  return <h1 className="text-xl text-arkw-el-6 p-4">Home</h1>;
-}
