+++ packages/admin/src/domains/integrations/components/integration-list-row.tsx
@@ -12,7 +12,7 @@ import { Dropdown } from '@/components/ui/dropdown-menu';
 
 import { capitalizeFirstLetter } from '@/lib/string';
 
-import { connectIntegrationByAPIKey } from '@/app/integrations/actions';
+import { connectIntegrationByAPIKey } from '@/app/(dashboard)/integrations/actions';
 
 import { IntegrationConnectDialog } from './integration-connect-dialog';
 import { ReferenceDialog } from './reference-dialog';
